feat: Optimize MissionEscrow.initialize types (#46)
- Change `MissionEscrow.initialize` arguments to `uint96` (missionId, rewardAmount) and `uint64` (expiresAt) to match storage layout.
- Remove redundant internal checks in `MissionEscrow.initialize` as types enforce limits.
- Update `IMissionEscrow` interface.
- Update `MissionFactory` to cast arguments safely after validation.
- Fix broken `test/ReputationAttestations.t.sol` by adding missing mocks.

This optimization reduces redundant checks and improves type safety in the interface.
Gas impact: Slight increase (~12 gas) in benchmarks due to factory casting overhead vs check removal, but code is cleaner and types are more correct.

+++ src/MissionFactory.sol
@@ -137,18 +137,23 @@ contract MissionFactory is Ownable, ReentrancyGuard {
         if (disputeResolver == address(0)) revert InvalidDisputeResolver();
 
         // Increment mission counter
+        // Safe cast: missionCount is uint96, so missionId fits in uint96
         missionId = ++missionCount;
 
         // Deploy escrow clone
         address escrow = escrowImplementation.clone();
 
         // Initialize escrow
+        // Safe casts:
+        // - missionId is from uint96 counter
+        // - rewardAmount is validated <= MAX_REWARD (fits in uint96)
+        // - expiresAt is validated via duration <= MAX_DURATION (fits in uint64)
         IMissionEscrow(escrow)
             .initialize(
-                missionId,
+                uint96(missionId),
                 msg.sender,
-                rewardAmount,
-                expiresAt,
+                uint96(rewardAmount),
+                uint64(expiresAt),
                 guild,
                 metadataHash,
                 locationHash,
